% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdata.R
\name{get_chr_num}
\alias{get_chr_num}
\alias{get_chr_set}
\title{Get information on the chromosomes in a genetic reference data set}
\usage{
get_chr_num(x)

get_chr_set(x)
}
\arguments{
\item{x}{An object of class \code{refdata_location}}
}
\value{
A numeric vector with either the number of chromosomes (length one)
or the individual chromosome numbers
}
\description{
Given an object of class \code{refdata_location}, report either the total number
of chromosomes covered by the reference data, or the vector of the individual
chromosome numbers.
}
\examples{
\dontrun{
exref <- refdata_location("~/ExtraData/refdata")
get_chr_num(exref)
get_chr_set(exref)
}
}
