% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdata.R
\name{get_variants}
\alias{get_variants}
\alias{get_chrmat}
\title{Load genetic reference files}
\usage{
get_variants(x)

get_chrmat(x, chr)
}
\arguments{
\item{x}{An object of class \code{refdata_location}}

\item{chr}{Integer or character, specifying the chromosome for which to load
the LD-structure matrix}
}
\value{
A character string with the path of the specified file
}
\description{
Given an object of class \code{refdata_location}, load the specified reference data
files (variant list, LD-structure matrices) from disk into R.
}
\examples{
\dontrun{
exref <- refdata_location("~/ExtraData/refdata")
get_variants(exref)
get_chrmat(exref, chr = 1)
}
}
