% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logit}
\alias{logit}
\alias{sigmoid}
\title{Logit and inverse logit as matrix transformations}
\usage{
logit(x, d = dim(x), dn = dimnames(x))

sigmoid(x, d = dim(x), dn = dimnames(x))
}
\arguments{
\item{x}{Designed for a numerical matrix of proportions, but should work
for vectors and general arrays, too.}

\item{d}{The dimension of the result; default is the dimension of the input \code{x}.
Set to \code{NULL} to return a vector.}

\item{dn}{A list of dimension names for the result; default are the dimension
names of the input \code{x}. Set to \code{NULL} to suppress dimension names.}
}
\value{
A vector, matrix or array, depending on the input format and the
specification for argument \code{d}.
}
\description{
These functions calculate the logit and inverse logit (sigmoid function)
for each element of a matrix of proportions, preserving dimension names if
defined.
}
\details{
No checks of parameters \code{d} and \code{dn} are performed.
}
