% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{p2z}
\alias{p2z}
\alias{z2p}
\title{Convert between p-values and test statistics}
\usage{
p2z(p)

z2p(z)
}
\arguments{
\item{p}{A vector of p-values on the -log10-scale (i.e. zero to infinity)}

\item{z}{A vector of test statistics from a standard normal null distribution}
}
\value{
A vector of test statistics or p-values, respectively
}
\description{
Helper functions for converting between two-sided p-values on the -log10-scale
and positive test statistics
}
\examples{
p2z( -log10(0.05) )
z2p( 1.96 )
}
