% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prop_ci}
\alias{prop_ci}
\title{Exact binomial confidence intervals for a series of experiments}
\usage{
prop_ci(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{A vector that contains the number of successes for each experiment}

\item{n}{A vector that contains the number of trials per experiment}

\item{conf.level}{The desired confidence level; default 0.95}
}
\value{
A matrix with one row per experiment and two columns, holding the
lower and upper confidence limit
}
\description{
Given a series of binomial experiments, summarized as a vector of
successes and a vector of trials, this calculates exact confidence
intervals slightly faster than a wrapper for base-R \code{binom.test}
}
\details{
A simple pass-through wrapper for function \code{binom.confint} in
package \code{binom}.
}
\examples{
x <- c(0, 10, 240)
n <- c(5, 100, 500)
prop_ci(x, n)
}
\seealso{
\code{\link[binom]{binom.confint}}
}
