% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseSmooth2d.R
\name{sparseSmooth2d}
\alias{sparseSmooth2d}
\title{Smoothing of values on a 2D grid}
\usage{
sparseSmooth2d(values, weights, smooth_param = c(10, 10), weight_thres = 0.001)
}
\arguments{
\item{values}{an n x p matrix of values on a grid}

\item{weights}{a matrix of weights, with the same dimensions}

\item{smooth_param}{a vector of length 2, which probably holds the a smoothing
parameter for rows / columns}

\item{weight_thresh}{minimal value for weights, smaller values are replaced by
this threshold (default 0.001)}
}
\value{
A matrix of smoothed values
}
\description{
This implements an algorithm that accepts a 2d-grid of values (a matrix of
proportions for the package-specific use case) and smoothes them based on
some facny differencing.
}
\examples{
## Small example, reasonably easy to check
vm <- rbind( c(1, 2, 3, 2, 5), c(2, 2, 3, 1, 4),  c(1, 4, 2, 1, 3), c(6, 3, 1, 2, 2) )
wm <- matrix(1, nrow = 4, ncol = 5)
sparseSmooth2d(vm, wm)

## A bigger example (close to use-case size)
nr <- 30
nc <- 100
set.seed(1)
vm <- matrix(rnorm(nr*nc), nrow = nr)
wm <- matrix(runif(nr*nc), nrow = nr)
sm <- sparseSmooth2d(vm, wm)
}
