% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which_col}
\alias{which_col}
\title{Extract locations of non-zero rows in the column of sparse matrix}
\usage{
which_col(x, j)
}
\arguments{
\item{x}{A sparse matrix}

\item{j}{The column index of interest}
}
\value{
A numerical vector no longer than the number of rows of x
}
\description{
Extract the row-indices of non-zero entries for a given column of a
sparse matrix in compressed column format (i.e. the R default)
}
\details{
Functionally the same as \code{which(x[,j])}, but more than 1E3 times
faster
}
\examples{
mm <- cbind( c(1, 0, 0, 0), c(0, 1, 0, 1), c(1, 0, 1, 0) )
mm <- Matrix::Matrix(mm, sparse = TRUE)

which_col(mm, 1)
which_col(mm, 2)
which_col(mm, 3)
}
