% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdr_pleio.R
\name{cfdr_pleio}
\alias{cfdr_pleio}
\title{A class for calculating conditional & conjunctional fdr for pleiotropy analysis}
\description{
A class that captures all data for calculating the conditional
and conjuncational fdr for pleiotropy analysis, and manages the necessary
analysis flow
}
\examples{
cfdr_pleio$new()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{trait1_data}}{Genetic data for the first trait}

\item{\code{trait2_data}}{Genetic data for the second trait}

\item{\code{traitnames}}{Character vector of length two, the names of the traits}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{cfdr_pleio$new()}}
\item \href{#method-clone}{\code{cfdr_pleio$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize an empty
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cfdr_pleio$new(trait1 = NA, trait2 = NA, traitnames)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trait1}}{Genetic data for the first trait}

\item{\code{trait2}}{Genetic data for the second trait}

\item{\code{traitnames}}{Character vector of length two, the names of the traits}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing special
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cfdr_pleio$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
